[FREELDR] - Use the new realmode code for PxeCallApi, SoftReboot and ChainLoadBiosBootSectorCode - Remove o386cpi.S from build, the functions are not used - Use constants for function ids - Fix a warning
svn path=/trunk/; revision=52284

+++ reactos/boot/freeldr/freeldr/arch/i386/i386idt.c
@@ -3,7 +3,7 @@
 
 
 KIDTENTRY DECLSPEC_ALIGN(4) i386Idt[32];
-KDESCRIPTOR i386IdtDescriptor = {0, 255, i386Idt};
+KDESCRIPTOR i386IdtDescriptor = {0, 255, (ULONG)i386Idt};
 
 static
 void
